emoji-madness: add reworked plugin
Since the last version was reverted because the use of :shortcodes:
would bleed into other players chat, i thought that i could rework this
to work like the EmojiScape external plugin where normal words or
phrases will be used as triggers for the emojis. In this way, there will
be no :codes: typed and i have also removed weird phrases like `small
medium black box` from the emojis and subsequently trigger phrases.

Typing `man in suit` will either append the emoji to the phrase or
replace it entirely for example.