Update MLM Upper Level Markers to 1.2.3
Replaced a config field for Contour Timer Markers to an enum type, allowing more marker types in the future.

This is mostly for adding a way of hiding the tile markers for people who use the timers and want less visual clutter.

Also adds a toggle for showing the text outline for the timers.