Update wiki-sync to 2.0
This is a core rewrite to the plugin and replaces the change-based approach into something that periodically just dumps all the relevant data for the user, computes the delta versus what has already been sent, and sends that to the server.

This gives us automatic retries (no changes get dropped if the server 500s), and less confusing logic if player name or profile change. It's also about 65% less code.