Refreshed the Organized Crime Plugin
* Original creator has been added as collaborator to the repository.

Plugin creator didn't respond for over a year to pull requests and the plugin contains a nasty bug that stops the plugin from working altogether. It's been fixed in this repo, and I've fixed some other problems with the plugin documented in README.md. Planning to add more features soon.