Fix bug when loading specific images
Images for a monster are being loaded using the monster's name. There were a few instances where the naming had inconsistent casing between the monster name and the image filename, these are the ones that were failing to load.

I have made all the image filenames lowercased, then before we attempt to load an image using the monster name we convert the name to lowercase.

Not sure why the images were able to load when ran in a local environment but not in live.