Update Twitch Live Loadout Plugin to v1.0.3
This updates the plugin to be compatible with the recent hitsplat API changes. The plugin uses the hitsplat types to categorize different types of damages and show them separately to Twitch viewers. 

With the deprecation of `Hitsplat.HitsplatType` we now moved to an integer based approach where we keep track of the hitsplat types ourselves to ensure no breaking changes in the future.