Add item-retrieval-warning
This plugin keeps track of whether a player has items stored in an item retrieval service or not and if they have, it warns the player when their health drops below a user-specified percentage by sending a notification, flashing the in-game screen red and displaying overlay message.
Warnings will not trigger in areas of the game where the items in an item retrieval service are not at risk. (Minigames, cox etc).