Update profit-tracker to the latest runelite.
The original author appears to have abandoned this plugin (https://github.com/runelite/plugin-hub/pull/3257). The plugin is not perfect, but I still found it very useful. I updated the links to point to my repo which has the fixes (including the open PR from the original repo that I verified locally) to get the plugin working in the latest runelite.

If/when the author decides to merge in the PRs and revive their repo, we can update the references back to `https://github.com/nofatigue/runelite-profit-tracker.git`