wilderness-clan-callouts
The "WildernessClanCallouts" plugin is a RuneScape plugin that allows players to quickly and easily call out the names of enemies in the Wilderness with a customizable prefix and hotkey. The plugin also has the option to shorten the enemy's name to 3 characters for faster typing. This plugin is useful for players who frequently participate in player vs player combat in the Wilderness and want to quickly and easily communicate the names of their enemies to their team.