Added plugin temple-trek-bog-helper
# Runelite Temple Trek Bog Helper
Highlights firm bog ground as you find them

During the bog event in Temple Trekking, as you discover tiles that are "firm" they will be highlighted automatically for you. These highlights will only persist for that single event and be reset when the bog puzzle resets.  

There is also a full on solver which automatically highlights all firm tiles without having to poke around in the bog.