update guild-spam-filter
* Update guild-spam-filter

* Revert

* Updated Guild Spam Filter

We now support PK broadcasts and we can now decide ourselves which collection log to filter out

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Update guild-spam-filter

* Updated Clan Spam Filter

Fixed bugs which was coursed by Jagex messing broadcast messages. We now always remove a broadcast if the setting is set to true and we couldn't find the right indexes to search by
Also I updated to latest RuneLite version

* Updated Clan Spam Filter

It is now possible through CSV to select which player names should always be allowed for broadcasts. 
Also updated to latest version of RL.

* Update guild-spam-filter