runewatch: update to v2.4.0
This new version of the RuneWatch plugin separates the watchlist json file and the plugin into their seperate branches. 

The old process contained commits both from the watchlist and plugin making it difficult to collab on just plugin changes.

The watchlist & updater moved to the `runewatch-updater` branch and the code is now pointing at the new URL for the json list.