Adjust popinbutton2 and tooltip
I moved popinbutton2 out of the config section and into popout section and that seems to display better.

Added (can cause weirdness) to tooltip about Always on Top after briefly experiencing a strange windowing issue on X11 that I can't seem to recreate outside of a single day. Not sure if it was related, disclaimer just in case.