Update displaynamedisguiser
Added a larger list of default names to the Config
Reorganized Config and Added Sections
Fixed an issue with usernames that had a hyphen?
Experimental Fix for a conflict with the Global Chat plugin
Fixed an instance of a missing icon
Partial Fix for improperly formatted config. It now ignores entries that are null, empty, a single space.