Update ectoplasmator-reminder
Changed the structure of the plugin to fetch a list of NPC names from another branch, and compare the name of the NPC targets (Instead of IDs). 

This is to set up a structure for future updates, where only the text file on the other branch needs to be updated with new Spectral NPCs, and no changes need to be made to the function or structure of the code on main.