update ectoplasmator-reminder (#7326)
* Uploading new plugin details - ectoplasmator reminder

* changed file title to fix merge error

* Update ectoplasmator-reminder

Updated commit version
Was using @ConfigGroup("example") - cause an issue, @ConfigGroup has been renamed

* Changed commit version - new feature, added config to hide overlay for bosses

* Changed capitalizations in config menu, images, and readme

* Update ectoplasmator-reminder (Varlamore Part 2 Creatures)

Added Varlamore Part 2 Creatures. 
Frost Nagua & Amoxilatl boss

* Update ectoplasmator-reminder

Changed the structure of the plugin to fetch a list of NPC names from another branch, and compare the name of the NPC targets (Instead of IDs). 

This is to set up a structure for future updates, where only the text file on the other branch needs to be updated with new Spectral NPCs, and no changes need to be made to the function or structure of the code on main.

* Update ectoplasmator-reminder

Changed URL Reader to Async OkHTTP calls

* Changed OkHTTP calls from Gradle import to @Inject

Changed OkHTTP calls from using a gradle / maven import, to an @Inject to call built in plugins.

First time using OkHTTP so I did not know what I was doing.

* Changed class modifiers on SpectralCreatures for OkHTTPClient to work

Removed abstract class modifier from SpectralCreatures class so the @ Inject for OkHTTPClient could work. 

While the class was abstract, my IDE kept telling me to make everything static. 

I had to go change how SpectralCreatures is called in the plugin using an @ Inject. 

This seems to have fixed all the issues on my side for testing, now we will see if the automatic GitHub checks pass

* Update ectoplasmator-reminder (Logger)

Changed console prints to logger

* Changing logging to Slf4j

See
https://github.com/runelite/plugin-hub/pull/7326#issuecomment-2644563299

For code review suggesting this change