update droptracker to v3.9 (#8033)
* v3.9

Re-implement api-based tracking (optional) to combat the inevitable outcome of "public" webhooks.

* Increment version to 3.9, add KC to drop & pb submissions

* Bugfix

Fixed an error preventing the plugin from starting

* Add Yama to longNpcs list

Added yama to the list of NPCs that take an unusually long period of time between the PB/kill time message and the loot event.

* Update droptracker

Call sendChatMessage on the client thread