add scrollboxcounter (#8411)
* Create scrollboxcounter

* added chat message

A chat message is displayed if a scroll box is picked up from the ground: "holding 2/3 scroll boxes (hard)"

* fixed chat message send on inventory update

Chat message was only sent on item pickup. Now it also shows on scroll box "appearing" in your inventory (for example, when woodcutting). Also checks if you withdraw from a bank interface to not send the chat msg

* removed deprecated api methods

* added config option for chat message