Update royal-titans-damage-tracker-plugin
Added flags and logic to properly handle resetting of the damage tracker.

The previous bug of the tracker resetting after the titans were killed still persisted. The issue was the checkAreaExit() method would assume the player left the titans area. Since the reset delay only activated when the player looted the titan, the time between the titans dying (NPC IDs becoming null) and the player looting the titans (triggering the chat message to reset the tracker) was treated as a player leaving the titans area (encounter ended without needed to activate the reset timer).