update royal-titans-damage-tracker-plugin (#9157)
Added the option for players to include burn damage into the damage trackers contribution and drop rate calculation. 

It is off by default but can be toggled on if the player wants to include burn in the tracker. Seeing as burn is a damage over time effect with the hitsplats isMine() and isOther() both being false, it would be difficult to attribute the damage to the correct player in the event both players are applying burn. But with this, the player has options to play around with.