Create Flux
This is a plugin I have written for my clan that was inspired by the clan events plugin. It is almost entirely just UI and swing components being updated in the plugin panel. There are two external READ only calls. One is to the google sheets API which is where our clan admins control the plugin config (only a few values for dynamic events). One is to Wiseoldman.net. No data is being written outside of the config file. 

There is a login message sent, clan ranks checked for my specific clan only, and some interaction with the Overlay class for automatic event password displaying for active events in WOM. Outside of that, it is only Javaswing in the PluginPanel class using the Card layout to display information.