Plugin Submission: Prometheus Skill Exporter
This plugin can be used by the player to submit their skills every X minute(s). It exposes a port on the player's machine so that they may add it as a target into prometheus. I will be adding grafana dashboard for this shortly. I plan to extend this to bank value tracking, collection logs, boss drops, and much more.